Debug

+++ transformer_f0_wav/utils.py
@@ -85,7 +85,7 @@ def worldSynthesize(wav_path,target_sr=44100,hop_length=512,fft_size=2048,f0_in 
     f0, t = pyworld.dio(wav.astype(np.double),fs=target_sr, frame_period=1000 * hop_length/target_sr)
     f0 = pyworld.stonemask(wav.astype(np.double), f0, t, target_sr)
     if f0 is not None:
-       f0 = f0_in
+       f0 = f0_in.astype(np.double)
     ap = pyworld.d4c(wav.astype(np.double), f0, t, target_sr, fft_size=fft_size)
     sp = pyworld.cheaptrick(wav.astype(np.double), f0, t, target_sr, fft_size=fft_size)
     synthesized = pyworld.synthesize(f0, sp, ap, target_sr, frame_period=1000 * hop_length/target_sr)
